requirements.txt
openai
streamlit
python-dotenv
langchain
langchain-openai
faiss-cpu
tiktoken